refactor(server): extract execution-phase modules from QueryExecutor (#851 PR2b)
Split the 1,264-line executor.ts god object by extracting five
single-responsibility execution-phase modules under src/server/execution/:

- ModeRouter            (mode-router.ts)            — resolve + validate query mode
- QueryResultCache      (query-result-cache.ts)     — cache key/lookup/store + events
- FilterCachePreloader  (filter-cache-preloader.ts) — pre-build filter SQL for dedup
- buildAnnotations      (annotation-builder.ts)     — UI annotation metadata (pure)
- postProcessResultRows (result-post-processor.ts)  — time-dim normalise + gap fill

QueryExecutor now composes these (modeRouter/resultCache/filterCachePreloader
fields; annotation + post-processing are pure functions). The inline cache-get
block in execute() collapses to generateKey + lookup; cacheResult/generateAnnotations/
preloadFilterCache/resolveQueryMode/validateQueryForMode removed. executor.ts
1,264 → 941 lines.

Behaviour-preserving: typecheck clean; 2348 server tests pass (postgres).
